refactor: web과 cli의 폴더 분리
